Add resource importer for metadata
This allows for audio files to be imported as a metadata object in editor, however, since one file cannot be imported in 2 ways, this will also limit that file from being used as a audio stream itself. This feature is mostly intended for use by other importers, or for looking up the output of metadata certain files in the editor.